feat: Profile ViewModel (#22)
* [WIP] Profile ViewModel and QR fetch
show profile, show registered events, fetch QR, update profile

* (fix) added @Module annotation

* feat: Added Profile and UpdateProfile ViewModels

Added Profile and Update Profile ViewModels to fetch QR code, show/update Profile, show Registered Events

* fix: removed such imports to pass ktlint

* Combined ViewModels

* feat: ProfileViewModel

* fix(profile): Removed UpdateUtils

* Moved constants to FestConstants.kt

* Wrote Comments

* fix: Bind ViewModel

+++ core/src/main/java/edu/nitt/delta/core/api/FestApiInterface.kt
@@ -115,7 +115,7 @@ interface FestApiInterface {
   /** Get User details */
   @GET(Routes.USER_DETAILS)
   suspend fun getUserDetails(
-    @Query("user_id") id: Int,
+    @Query("user_id") id: Long,
     @Query("user_token") token: String
   ): UserDetailsResponse
 
@@ -123,7 +123,7 @@ interface FestApiInterface {
   @FormUrlEncoded
   @POST(Routes.USER_UPDATE_DETAILS)
   suspend fun updateDetails(
-    @Field("user_id") id: Int,
+    @Field("user_id") id: Long,
     @Field("user_token") token: String,
     @FieldMap parameters: Map<String, String>
   ): UpdateDetailsResponse
